# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Livestock Price Board API
  version: 0.1.0-phase1
  summary: Phase 1 contract. Auth, farm and herd records, price board, admin verification and reference prices.
  description: |
    Stack-agnostic contract for the Phase 1 build (weeks 3 to 8). NestJS or Laravel
    can implement it unchanged. Phase 2 (listings, offers, deals) and Phase 3
    (shipments) will be added as separate tag groups without breaking these paths.

    Conventions
    - All responses are JSON. Errors use RFC 9457 `application/problem+json`.
    - Every mutating request from a mobile client carries `Idempotency-Key` so an
      offline queue can retry safely. Same key + same body returns the original
      response; same key + different body returns 422.
    - Lists are cursor paginated: `?cursor=&limit=` in, `next_cursor` out.
    - Money is a decimal string in Philippine pesos, e.g. "185.50". Never a float.
    - Weights are decimal strings in kilograms.
    - Timestamps are RFC 3339 UTC. Dates are `YYYY-MM-DD`.
    - Locations are PSGC codes (text). The client never sends free-text place names.
    - Every price returned carries `source`, `location_code` and `sample_count`.
      A client must display all three. A bare number is a defect.
    - Records that can be edited offline (farms, lots) carry `version`. PATCH must
      send `If-Match: <version>`; a stale version returns 409 with the server copy.

    Public board
...
      operations; nothing else changes.
  contact:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /deals/{deal_id}/deposit:
    get:
      tags: [Payments]
      operationId: getDeposit
      summary: The booking deposit on a deal
      description: Farmer, buyer and admins. checkout_url is returned to the buyer only, while the deposit is pending.
      parameters:
        - name: deal_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Deposit' }
        '404': { $ref: '#/components/responses/NotFound' }
  /deals/{deal_id}/deposit/checkout:
    post:
      tags: [Payments]
      operationId: refreshDepositCheckout
      summary: Buyer asks for a fresh checkout link
      description: Used when the first link could not be created or was closed. Same amount, same deadline.
      parameters:
        - name: deal_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Deposit' }
        '409':
          description: Deposit is not pending.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /me/payout-account:
    get:
      tags: [Payments]
      operationId: getMyPayoutAccount
      summary: Where I get paid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PayoutAccount' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Payments]
      operationId: putMyPayoutAccount
      summary: Set my GCash number or bank account
      description: Farmer role. A one-peso test transfer verifies the account; released deposits go here.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PayoutAccountInput' }
      responses:
        '200':
          description: Saved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PayoutAccount' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /payments/webhook:
    post:
      tags: [Payments]
      operationId: paymentsWebhook
      summary: Provider callbacks (checkout paid, transfer succeeded or failed)
      description: Unauthenticated; the body is verified with the provider's signature header. Events are stored once by id and replays return 200 without side effects.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: Recorded. }
        '400': { description: Bad signature or malformed event. }
  /admin/deposits:
    get:
      tags: [Payments]
      operationId: adminListDeposits
      summary: Deposits, newest first
      parameters:
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/DepositStatus' }
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items, totals]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Deposit' }
                  totals: { $ref: '#/components/schemas/DepositTotals' }
  # ------------------------------------------------------------------ Sync
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access token, 15 minute lifetime. Claims: `sub` (user id), `roles` (array),
        `ver` (verification status), `jti`. Role checks happen server-side on every
        request; the client uses the claims only to choose which screens to show.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Client-generated UUID. Stored 48 hours.
      schema: { type: string, format: uuid }
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
...
  responses:
...
    Forbidden:
      description: Role or ownership check failed.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Not found, or not visible to this user.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri, example: 'https://api.example.ph/problems/validation' }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
...
    Money:
      type: string
      pattern: '^[0-9]+\.[0-9]{2}$'
      description: Philippine pesos as a decimal string.
      example: '185.50'
    MoneyNullable:
      type: [string, 'null']
      pattern: '^[0-9]+\.[0-9]{2}$'
...
    DepositStatus:
      type: string
      enum: [pending, paid, lapsed, released, forfeited, refunded]
    DepositSummary:
      type: object
      required: [id, status, amount, expires_at]
      properties:
        id: { type: string, format: uuid }
        status: { $ref: '#/components/schemas/DepositStatus' }
        amount: { $ref: '#/components/schemas/Money' }
        expires_at: { type: string, format: date-time, description: 'Pay-by time while pending.' }
        paid_at: { type: [string, 'null'], format: date-time }
        closed_at: { type: [string, 'null'], format: date-time }
    Deposit:
      allOf:
        - $ref: '#/components/schemas/DepositSummary'
        - type: object
          required: [deal_id, provider, created_at]
          properties:
            deal_id: { type: string, format: uuid }
            buyer_id: { type: string, format: uuid }
            farmer_id: { type: string, format: uuid }
            buyer_name: { type: string }
            farmer_name: { type: string }
            provider: { type: string }
            checkout_url: { type: [string, 'null'], description: 'Buyer only, while pending.' }
            payment_method: { type: [string, 'null'] }
            fee: { $ref: '#/components/schemas/MoneyNullable' }
            net: { $ref: '#/components/schemas/MoneyNullable' }
            transfer_status: { type: [string, 'null'], enum: [pending, succeeded, failed, null] }
            failure_reason: { type: [string, 'null'] }
            created_at: { type: string, format: date-time }
    DepositTotals:
      type: object
      properties:
        held: { $ref: '#/components/schemas/Money', description: 'Sum of net over paid deposits. What the wallet must hold.' }
        released: { $ref: '#/components/schemas/Money' }
        forfeited: { $ref: '#/components/schemas/Money' }
        refunded: { $ref: '#/components/schemas/Money' }
        pending_count: { type: integer }
    PayoutAccountInput:
      type: object
      required: [kind, account_no, account_name]
      properties:
        kind: { type: string, enum: [gcash, bank] }
        account_no: { type: string, minLength: 6, maxLength: 34, description: 'GCash mobile number or bank account number.' }
        account_name: { type: string, minLength: 2, maxLength: 120, description: Must match the name on the ID. }
        bank_code: { type: [string, 'null'], maxLength: 20, description: 'Required for bank.' }
    PayoutAccount:
      allOf:
        - $ref: '#/components/schemas/PayoutAccountInput'
        - type: object
          required: [verified, updated_at]
          properties:
            account_no_masked: { type: string }
            verified: { type: boolean, description: 'True once the one-peso test transfer succeeded.' }
            updated_at: { type: string, format: date-time }
...